fix: keep flashing PACMAN when player buffers an input
Before this commit, PACMAN stops flashing when player
holds down an arrow key before arriving at a corner,
which was useful as this let the player know that their
input is being correctly buffered, so that the player
knows that they will successfully corner the incoming
corner, of course as long as they continue to correctly
buffer the input.

But this can look odd when the player buffers the input
well well in advance of the incoming corner, so this
fix is here to make PACMAN's animations look less odd.