refactor: move game memory allocation to constructor
Moves memory allocation from happening in game->run,
to happening in the constructor of a Game, following
the RAII pattern for memory acquisition.

So, game->run now also doesn't take in a Level to run:

The level that the game runs is now specified in the
game's constructor, as opposed to being specified in
game->run's arguments. And if the game's constructor
receives no specific level to load, it will load the
default level.